toggle switch widgets worked, refactor
