+++ tabulous/_colormap.py
@@ -18,7 +18,7 @@ _DEFAULT_MAX = "#FF696B"
 
 def exec_colormap_dialog(ds: pd.Series, parent=None):
     """Open a dialog to define a colormap for a series."""
-    from tabulous._qt._color_edit import ColorEdit
+    from tabulous._magicgui import ColorEdit
     from magicgui.widgets import Dialog, LineEdit, Container
 
     dtype = ds.dtype
