+++ tabulous/_magicgui/_color_edit.py
